Include relevant papers. Modify gitignore
Read through the papers relevant to the given project and understand all
the steps required for implementing the same.
Include the relevant papers that I read in the repo in GitHub.

Modify gitignore to ignore PyCharm, Jupyter Notebook stuff, MATLAB,
Octave and Vim.

Downloaded the dataset of 400 images for the project.
Modified .gitignore to ignore the datasets so that they are not pushed
to the GitHub repository